added real video for factories
