+++ apps/afisha/factories/events.py
@@ -1,3 +1,4 @@
+import random
 from zoneinfo import ZoneInfo
 
 import factory
@@ -5,6 +6,7 @@ from django.conf import settings
 
 from apps.afisha.models import CommonEvent, Event
 from apps.core.decorators import restrict_factory
+from apps.core.utils import get_video_in_channel
 
 
 @restrict_factory(general=(CommonEvent,))
@@ -12,11 +14,12 @@ class EventFactory(factory.django.DjangoModelFactory):
     """Create Event.
 
     Parameters:
-    1.`date_time_in_three_hours=True`: create event at random time but not
+    1. `date_time_in_three_hours=True`: create event at random time but not
     more than 3 hours from now.
     2. `masterclass=True`: bind event with random `master class`
     3. `reading=True`: bind event with random `reading`
     4. `performance=True`: bind event with random `performance`
+    5. `add_real_video=True`: create object with real link to youtube video`.
     """
 
     class Meta:
@@ -44,6 +47,9 @@ class EventFactory(factory.django.DjangoModelFactory):
             common_event=factory.LazyFunction(
                 lambda: CommonEvent.objects.exclude(performance__isnull=True).order_by("?").first()
             ),
+        )
+        add_real_video = factory.Trait(
+            url=factory.LazyFunction(lambda: random.choice(get_video_in_channel())),
         )
 
     date_time = factory.Faker(
