+++ apps/library/factories/play.py
@@ -2,7 +2,7 @@ import factory
 from faker import Faker
 
 from apps.core.decorators import restrict_factory
-from apps.core.utils import slugify
+from apps.core.utils import get_video_in_channel, slugify
 from apps.info.models import Festival
 from apps.library.models import Play, ProgramType
 
@@ -34,6 +34,11 @@ class PlayFactory(factory.django.DjangoModelFactory):
     url_download = factory.django.FileField()
     published = factory.Faker("boolean", chance_of_getting_true=50)
     other_play = False
+
+    class Params:
+        add_real_video = factory.Trait(
+            url_reading=factory.Iterator(get_video_in_channel, cycle=False),
+        )
 
     @factory.lazy_attribute
     def url_reading(self):
